some norm make bonus files with bonus ft for bonus compilation leaks management for esc keypad but not for the rest
