+++ libft/srcs/ft_strndup.c
