+++ srcs/display/minimap.c
@@ -3,22 +3,14 @@
 /*                                                        :::      ::::::::   */
 /*   minimap.c                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: hgeffroy <hgeffroy@student.42.fr>          +#+  +:+       +#+        */
+/*   By: xcharra <xcharra@student.42lyon.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/26 11:00:47 by hgeffroy          #+#    #+#             */
-/*   Updated: 2023/11/02 14:19:33 by hgeffroy         ###   ########.fr       */
+/*   Updated: 2023/11/03 15:25:29 by xcharra          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "cub3D.h"
-
-void	my_mlx_pixel_put(t_img *img, int x, int y, int color)
-{
-	char	*dst;
-
-	dst = img->addr + (y * img->line_len + x * (img->bits_per_pixel / 8));
-	*(unsigned int *)dst = color;
-}
 
 void	draw_tile(t_data *cub, float x, float y, float a, float b, int *centre)
 {
@@ -120,17 +112,6 @@ void	draw_player(t_data *cub)
 	centre[1] = 7 * TILE_SZ;
 	draw_fov(cub);
 	draw_circle(cub, centre, PLAYER_SZ, H_BLUE);
-}
-
-
-void	init_minimap(t_data *cub)
-{
-	cub->mlx = mlx_init();
-	cub->win = mlx_new_window(cub->mlx, SCREEN_WIDTH, SCREEN_HEIGHT, "cub3D");
-	cub->display.img = mlx_new_image(cub->mlx, SCREEN_WIDTH, SCREEN_HEIGHT);
-	cub->display.addr = mlx_get_data_addr(cub->display.img,
-			&cub->display.bits_per_pixel, &cub->display.line_len,
-			&cub->display.endian);
 }
 
 void	set_minimap(t_data *cub, t_minimap *minimap)
